Extracted "main" simulation function and options
